L5-Fixes. CartService отвязан от хранения данных в Cookies
