+++ Services/WebStore.Interfaces/Services/ICartStore.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using WebStore.Domain.Entities;
+
+namespace WebStore.Interfaces.Services
+{
+    public interface ICartStore
+    {
+        Cart Cart { get; set; }
+    }
+}
