+++ Services/WebStore.Services/Services/InCookies/InCookiesCartStore.cs
@@ -1,0 +1,55 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
+using Newtonsoft.Json;
+using WebStore.Domain.Entities;
+using WebStore.Interfaces.Services;
+
+namespace WebStore.Services.Services.InCookies
+{
+    public class InCookiesCartStore : ICartStore
+    {
+        private readonly IHttpContextAccessor _HttpContextAccessor;
+        private readonly string _CartName;
+
+        public InCookiesCartStore(IHttpContextAccessor HttpContextAccessor)
+        {
+            _HttpContextAccessor = HttpContextAccessor;
+
+            var user = HttpContextAccessor.HttpContext!.User;
+            var user_name = user.Identity!.IsAuthenticated ? $"-{user.Identity.Name}" : null;
+
+            _CartName = $"WebStore.Cart{user_name}";
+        }
+
+        public Cart Cart
+        {
+            get
+            {
+                var context = _HttpContextAccessor.HttpContext;
+                var cookies = context!.Response.Cookies;
+                var cart_cookie = context.Request.Cookies[_CartName];
+                if (cart_cookie is null)
+                {
+                    var cart = new Cart();
+                    cookies.Append(_CartName, JsonConvert.SerializeObject(cart));
+                    return cart;
+                }
+
+                //ReplaceCookies(cookies, cart_cookie);
+                return JsonConvert.DeserializeObject<Cart>(cart_cookie);
+            }
+            set => ReplaceCookies(_HttpContextAccessor.HttpContext!.Response.Cookies, JsonConvert.SerializeObject(value));
+        }
+
+        private void ReplaceCookies(IResponseCookies cookies, string cookie)
+        {
+            cookies.Delete(_CartName);
+            cookies.Append(_CartName, cookie);
+        }
+
+    }
+}
